requirements.txt
streamlit
crewai
python-dotenv
litellm
